eth-multichain-bridge-brownie/requirements.txt

bigchaindb_driver
eth-brownie
python-dotenv